backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "cauejlk", email = "cauejoaquimlk@gmail.com"},
]
dependencies = ["django>=5.2.5", "djangorestframework>=3.16.1", "django-cors-headers>=4.7.0"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
